juneau-app/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
aws-cdk-lib==2.185.0
aws-cdk.asset-awscli-v1==2.2.229
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==40.7.0
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
constructs==10.4.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jiter==0.9.0
jsii==1.110.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-core==0.3.48
langchain-text-splitters==0.3.7
langsmith==0.3.18
orjson==3.10.16
packaging==24.2
publication==0.0.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.3.0
uv==0.6.9
zstandard==0.23.0
